Changed project structure Also added code coverage.
